% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{act_method}
\alias{act_method}
\title{act_method}
\usage{
act_method(method, x)
}
\arguments{
\item{method}{String with the name of the activation method that will
be used. It must be one of \code{"step"}, \code{"sine"},
\code{"tangent"}, \code{"linear"}, \code{"relu"}, \code{"gelu"} or
\code{"swish"}. Anything else will raise an error.}

\item{x}{numeric on which the activation function will be applied. Typically, the sum of weighted weights.}
}
\value{
Value after applying the activation method.
}
\description{
Sets the activation method the user wants to use.
}
\keyword{internal}
