% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decission_tree.R
\name{comb}
\alias{comb}
\title{comb}
\usage{
comb(v, vp, k, combinations)
}
\arguments{
\item{v}{a}

\item{vp}{b}

\item{k}{c}

\item{combinations}{d}
}
\value{
Value of the new classified example.
}
\description{
Performes all possible combinations.
}
\keyword{K-Nearest}
\keyword{Neighbors,}
\keyword{classification,}
\keyword{distance.}
\keyword{knn,}
\keyword{supervised}
