% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db1rl}
\alias{db1rl}
\title{Test Database 1}
\format{
\subsection{\code{db1rl}}{

A data frame with 4 independent variables (first 4 columns, representing different line types).
The last column is the independent variable.
}
}
\usage{
db1rl
}
\description{
Test Database 1
}
\keyword{datasets}
