% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db2}
\alias{db2}
\title{Test Database 6}
\format{
\subsection{\code{db2}}{

A data frame with 3 independent variables (first 3 columns) and one independent variable (last column).
It has information about vehicles.
}
}
\usage{
db2
}
\description{
Test Database 6
}
\keyword{datasets}
