% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db3}
\alias{db3}
\title{Test Database 7}
\format{
\subsection{\code{db3}}{

A data frame with 3 independent variables (first 3 columns) and one independent variable (last column).
It has information about vehicles. Similar to db2 but a little bit more complex.
}
}
\usage{
db3
}
\description{
Test Database 7
}
\keyword{datasets}
