% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_flowers}
\alias{db_flowers}
\title{Test Database 5}
\format{
\subsection{\code{db_flowers}}{

A data frame representing features of flowers. It has 4 independent variables (first 4 columns) and one independent variable (last column).
}
}
\usage{
db_flowers
}
\description{
Test Database 5
}
\keyword{datasets}
