% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_per_and}
\alias{db_per_and}
\title{Test Database 2}
\format{
\subsection{\code{db_per_and}}{

A data frame with 3 independent variables (first 3 columns) and one independent variable (last column).
It represents a 3 input "AND" logic gate.
}
}
\usage{
db_per_and
}
\description{
Test Database 2
}
\keyword{datasets}
