% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_per_or}
\alias{db_per_or}
\title{Test Database 3}
\format{
\subsection{\code{db_per_or}}{

A data frame with 3 independent variables (first 3 columns) and one independent variable (last column).
It represents a 3 input "OR" logic gate.
}
}
\usage{
db_per_or
}
\description{
Test Database 3
}
\keyword{datasets}
