% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_per_xor}
\alias{db_per_xor}
\title{Test Database 4}
\format{
\subsection{\code{db_per_xor}}{

A data frame with 3 independent variables (first 3 columns) and one independent variable (last column).
It represents a 3 input "XOR" logic gate.
}
}
\usage{
db_per_xor
}
\description{
Test Database 4
}
\keyword{datasets}
