% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{db_tree_struct}
\alias{db_tree_struct}
\title{Test Database 8}
\format{
\subsection{\code{db_tree_struct}}{

Decision tree structure. output of the decision_tree() function "decision_tree(db2, "VehicleType", 4, "gini")"
}
}
\usage{
db_tree_struct
}
\description{
Test Database 8
}
\keyword{datasets}
