% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{perceptron}
\alias{perceptron}
\title{Perceptron}
\usage{
perceptron(
  training_data,
  to_clasify,
  activation_method,
  max_iter,
  learning_rate
)
}
\arguments{
\item{training_data}{hola}

\item{to_clasify}{hola}

\item{activation_method}{hola}

\item{max_iter}{hola}

\item{learning_rate}{hola}
}
\value{
mete los ejemplos perrrrrooooooo @examples
}
\description{
Applies k-nn algorithm to classify data.
}
\author{
Víctor Amador Padilla, \email{vamadorpadilla@edu.uah.es}
}
\keyword{hola}
