% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_regression.R
\name{polynomial_regression}
\alias{polynomial_regression}
\title{Multivariate Polynomial Regression}
\usage{
polynomial_regression(data, degree, details)
}
\arguments{
\item{data}{x*y data frame with already classified observations. Each column
represents a parameter of the values (independent variable). The last column
represents the classification value (dependent variable). Each row is a different observation.}

\item{degree}{Degree of the equations approximation.}

\item{details}{Boolean value. If it is set to "TRUE" multiple clarifications
and explanations are printed along the code}
}
\value{
vector with 2 values (If covariance = 0 the return vector will be empty):
* "R^2": Value that represents how well the regression line fits the given
values.
* "f(x)": Regression line equation.
}
\description{
This function calculates and plots the linear regression of a
given set of values. Being all of them independent values but one, that is the dependent value
}
\keyword{classification,}
\keyword{gain}
\keyword{information}
\keyword{learning,}
\keyword{linear}
\keyword{regression,}
\keyword{supervised}
