% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decission_tree.R
\name{print.tree_struct}
\alias{print.tree_struct}
\title{Print Tree Structure}
\usage{
\method{print}{tree_struct}(x, ...)
}
\arguments{
\item{x}{The tree structure.}

\item{...}{Extra useless parameters.}
}
\value{
nothing.
}
\description{
This function prints the structure of a tree, generated by the
\code{decision_tree} function.
}
\details{
It must receive a \code{tree_struct} data type.
}
\examples{
# example code
print(db_tree_struct)

}
\author{
Víctor Amador Padilla, \email{victor.amador@edu.uah.es}
}
